test(e2e): split ra-all and helpers by area
ra-all.spec.ts held 54 tests in one 125 KB file behind one describe, so
one early failure in serial mode took every later area with it. It is now
seven specs, one per area: keyboard, mouse, virtual media, USB gadget
config, display, TLS, factory reset. The page, agent, macro setup and log
capture they shared live in remote-agent/shared.ts; each spec registers
its session with registerSharedSession() and keeps serial mode. The
factory reset spec reboots the device into setup, so it runs in its own
project after the rest of the remote-agent project.

helpers.ts, 1,600 lines, is now e2e/helpers/ with device, hid, video,
auth, ssh and ota modules and an index that re-exports all of them, so
every existing import path still resolves.

Test bodies and helper bodies are moved, not edited. playwright test
--list reports the same 95 tests.
